+++ react/AffiliateAddPage.tsx
@@ -3,7 +3,8 @@ import React from 'react'
 import {
   Page,
   PageHeader,
-  PageTitle,
+  PageHeaderTitle,
+  PageHeaderTop,
   PageContent,
   createSystem,
   ToastProvider,
@@ -14,9 +15,7 @@ import { useIntl } from 'react-intl'
 import AffiliateForm from './components/admin/affiliates/form/AffiliateForm'
 import { messages } from './utils/messages'
 
-const [ThemeProvider] = createSystem({
-  key: 'affiliates-add',
-})
+const [ThemeProvider] = createSystem()
 
 const AffiliateAddPage: FC = () => {
   const intl = useIntl()
@@ -33,9 +32,11 @@ const AffiliateAddPage: FC = () => {
       <ToastProvider>
         <Page>
           <PageHeader onPopNavigation={handleBackAction}>
-            <PageTitle>
-              {intl.formatMessage(messages.addAffiliateTitle)}
-            </PageTitle>
+            <PageHeaderTop>
+              <PageHeaderTitle>
+                {intl.formatMessage(messages.addAffiliateTitle)}
+              </PageHeaderTitle>
+            </PageHeaderTop>
           </PageHeader>
           <PageContent>
             <AffiliateForm />