+++ react/AffiliateEditPage.tsx
@@ -3,7 +3,8 @@ import React from 'react'
 import {
   Page,
   PageHeader,
-  PageTitle,
+  PageHeaderTitle,
+  PageHeaderTop,
   PageContent,
   createSystem,
   ToastProvider,
@@ -17,9 +18,7 @@ import { messages } from './utils/messages'
 import AffiliateForm from './components/admin/affiliates/form/AffiliateForm'
 import GET_AFFILIATE from './graphql/getAffiliate.graphql'
 
-const [ThemeProvider] = createSystem({
-  key: 'affiliates-edit',
-})
+const [ThemeProvider] = createSystem()
 
 type AffiliateQueryReturn = {
   getAffiliate: Affiliate
@@ -58,9 +57,11 @@ const AffiliateEditPage: FC = () => {
       <ToastProvider>
         <Page>
           <PageHeader onPopNavigation={handleBackAction}>
-            <PageTitle>{`${intl.formatMessage(
-              messages.affiliateLabel
-            )}: ${affiliateId}`}</PageTitle>
+            <PageHeaderTop>
+              <PageHeaderTitle>{`${intl.formatMessage(
+                messages.affiliateLabel
+              )}: ${affiliateId}`}</PageHeaderTitle>
+            </PageHeaderTop>
           </PageHeader>
           <PageContent>
             <AffiliateForm affiliate={data?.getAffiliate} />