+++ react/components/admin/affiliates/form/AffiliateForm.tsx
@@ -1,5 +1,5 @@
 import { Box, Button, Flex, FlexSpacer, useToast } from '@vtex/admin-ui'
-import { Form, Formik } from 'formik'
+import { Form, useFormState, yupResolver } from '@vtex/admin-ui-form'
 import type { FC } from 'react'
 import React, { useCallback, useMemo } from 'react'
 import { useMutation } from 'react-apollo'
@@ -113,13 +113,21 @@ const AffiliateForm: FC<AffiliateFormProps> = ({ affiliate }) => {
         updateAffiliate({
           variables: {
             affiliateId: affiliate.id,
-            updateAffiliate: { ...values, isApproved: affiliate.isApproved },
+            updateAffiliate: {
+              ...values,
+              phone: values.phone?.toString(),
+              isApproved: affiliate.isApproved,
+            },
           },
         })
       } else {
         addAffiliate({
           variables: {
-            newAffiliate: { ...values, isApproved: false },
+            newAffiliate: {
+              ...values,
+              phone: values.phone?.toString(),
+              isApproved: false,
+            },
           },
         })
       }
@@ -127,29 +135,27 @@ const AffiliateForm: FC<AffiliateFormProps> = ({ affiliate }) => {
     [affiliate, updateAffiliate, addAffiliate]
   )
 
+  const form = useFormState({
+    resolver: yupResolver(VALIDATION_SCHEMAS(intl).affiliateForm),
+    defaultValues: initialValues,
+  })
+
   return (
     <Box csx={{ marginY: 16 }}>
-      <Formik
-        onSubmit={onSubmit}
-        initialValues={initialValues}
-        enableReinitialize
-        validationSchema={VALIDATION_SCHEMAS(intl).affiliateForm}
-      >
-        <Form>
-          <GeneralInfo />
-          <AddressInfo />
-          <MarketingInfo />
-          <Flex>
-            <FlexSpacer />
-            <Button
-              loading={editMutationLoading || addMutationLoading}
-              type="submit"
-            >
-              {intl.formatMessage(messages.saveLabel)}
-            </Button>
-          </Flex>
-        </Form>
-      </Formik>
+      <Form state={form} onSubmit={onSubmit}>
+        <GeneralInfo form={form} />
+        <AddressInfo form={form} />
+        <MarketingInfo form={form} />
+        <Flex>
+          <FlexSpacer />
+          <Button
+            loading={editMutationLoading || addMutationLoading}
+            type="submit"
+          >
+            {intl.formatMessage(messages.saveLabel)}
+          </Button>
+        </Flex>
+      </Form>
     </Box>
   )
 }