+++ react/components/admin/shared/TableActions.tsx
@@ -1,10 +1,4 @@
-import {
-  Menu,
-  MenuButton,
-  MenuItem,
-  MenuList,
-  useMenuState,
-} from '@vtex/admin-ui'
+import { Menu, MenuButton, MenuItem, useMenuState } from '@vtex/admin-ui'
 import React from 'react'
 
 import type { Action } from '../../../typings/tables'
@@ -34,27 +28,21 @@ const TableActions = ({ actions }: TableActionsProps) => {
   const state = useMenuState()
 
   return (
-    <Menu state={state} hideOnClick csx={Styles.Main}>
-      <MenuButton
-        display="actions"
-        variant="tertiary"
-        className="line-action-button"
-      />
-
-      <MenuList aria-label="actions" state={state} csx={Styles.Container}>
+    <>
+      <MenuButton variant="tertiary" state={state} />
+      <Menu aria-label="actions" state={state} csx={Styles.Container}>
         {actions.map((action) => {
           return (
             <MenuItem
               key={action.label}
               onClick={action.handleOnClick}
               icon={action.icon}
-            >
-              {action.label}
-            </MenuItem>
+              label={action.label}
+            />
           )
         })}
-      </MenuList>
-    </Menu>
+      </Menu>
+    </>
   )
 }